requirements.txt
flask==2.0.1
numpy==1.23.5
opencv-python-headless==4.5.1.48
pytesseract==0.3.9
easyocr==1.6.2
Pillow==9.0.1
ultralytics==8.0.20
gunicorn==20.1.0
werkzeug==2.0.1